Update TimeoutStressTest to ensure pool grows back to core after test.
Adds additional validation to the test such that it ensures that once
timeouts are restored to reasonable values that the pool grows back to
an expected core state.

Reduce duration of test to 1 minute.  As a default this test should
serve more as a sanity check to ensure connections don't leak.  In local
testing about 4000 channels are created within 1 minute which should be
enough time to find an obvious leak.

Add test for verifying PoolingOptions#setInitializationExecutor.
